requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.2.0
greenlet==2.0.2
idna==3.4
pathlib==1.0.1
psycopg2==2.9.7
python-telegram-bot==13.15
pytz==2023.3
requests==2.31.0
schedule==1.2.0
six==1.16.0
SQLAlchemy==2.0.20
tornado==6.1
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
